Moved Neo4j config and driver set-up into dedicated db folder in lib and updated imports where necessary.
